Fix food loaded from a saved step always having 0 nutrients and energy
There was a max imposed on the setter which was set after rather than before. It's only used as a quick and dirty way for creatures to know the quality of food so it doesn't make too much of a difference to strip the limiting functionality from the food itself.